+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-[not released] - 2014-08-22
+[0.0.3] - 2014-08-23
   * Added railtie to require controller extension instantly
+  * Added :all_actions matcher
+  * Configuration for authentication_validation_method
 
 [0.0.1] - 2014-08-21 Initial import
   * Includes plain and conditional restrictions